Move "nux-welcome" to layout folder.
