+++ client/layout/index.jsx
@@ -21,7 +21,7 @@ var MasterbarLoggedIn = require( 'layout/masterbar/logged-in' ),
 	PreferencesData = require( 'components/data/preferences-data' ),
 	EmailVerificationNotice = require( 'components/email-verification/email-verification-notice' ),
 	Welcome = require( 'my-sites/welcome/welcome' ),
-	WelcomeMessage = require( 'nux-welcome/welcome-message' ),
+	WelcomeMessage = require( 'layout/nux-welcome/welcome-message' ),
 	GuidesTours = require( 'guidestours' ),
 	analytics = require( 'lib/analytics' ),
 	config = require( 'config' ),
