+++ client/layout/nux-welcome/index.js
