+++ client/layout/nux-welcome/welcome-message.jsx
